solves polynomials using coefficients and x values
Each pair of lines will represent one problem. The first line will contain a list of integers which represent a set of coefficients to a polynomial expression. The order of the polynomial is n. The second line of text represents a sequence of values for x. For each pair of lines, your program should evaluate the polynomial for all the values of x and output the resulting values on a single line.